Scaleflex/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
requests>=2.8.1
blurhash-python>=1.0.2
djangorestframework>=3.11.0